small refactor for comments
